pyproject.toml
[tool.poetry]
name = "mangi"
version = "1.0.0"
description = "Telegram bot for the canteens of the TU Dresden"
authors = ["Nick Lehmann <nick@lehmann.sh>"]

[tool.poetry.dependencies]
python = ">=3.8"
requests = ">2"
beautifulsoup4 = ">4"
python-telegram-bot = ">12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
